own help done

+++ own_hlp.c
@@ -1,7 +1,7 @@
 #include "main.c"
 
 /**
- * own_help - print help
+ * help - print help
  * @args: arguments
  *
  * Return: 1 on success, 0 otherwise
@@ -15,7 +15,7 @@ int help(char **args)
 		"help",
 		"exit"
 	};
-	long unsigned int i = 0;
+	unsigned int i = 0;
 	(void)(**args);
 
 	printf("\n---help simple_shell---\n");
